Database backup/restore tool documentation added
