--- backend/restore-firestore/.vscode/settings.json
@@ -1,7 +1,0 @@
-{
-    "workbench.colorCustomizations": {
-        "activityBar.background": "#063148",
-        "titleBar.activeBackground": "#094465",
-        "titleBar.activeForeground": "#F7FCFE"
-    }
-}
